Improve crash when interpreter fails to increment
JavaScript's Number type can only represent integers up to 2^53. When
the Nock interpreter attempts to increment (with Nock operator 4) an
atom equal to 2^53, it will result in the same atom, which violates the
Nock specification. The interpreter will detect a loop and crash, but
there are two problems: (1) it's not crashing for the right reason, and
(2) it doesn't give helpful information about what went wrong. This
commit modifies the interpreter to check for this case and throw an
error.